+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+Updated
+-------
+- Preferring Bearer token over user token when configuring Connection session
+
 v9.6.0
 ------